refactor part 6, bug fixes part 1
